third_party/WebKit/LayoutTests/editing/pasteboard/paste-double-nested-blockquote-before-blockquote-expected.txt

CONSOLE MESSAGE: line 18: SUCCESS: pasting doubleNestedNodeToCopy DID NOT keep the newline within the blockquote.
This tests pasting a doubly nested blockquote before, and outside of, another blockquote. This test does not work outside of DRT.

line 4


line 1
line 2
line 3
line 4
line 5
line 6
